Extracting the eta functionality into a new file. This provides a way to use it in other places.
